Adding dice rolling functionality.
Added the rand crate to the backend for generating random numbers.
Implemented a generic DieRoller in rolls.rs to generate numbers according to the number of sides or the number of sides and quantity of dice; Includes a couple tests to confirm the results out to 100 rolls.
Added a new RollDice command to main.rs.
Added a new event listener to App.js for handling the result of the RollDice command being passed back to the frontend.
Moved the process of parsing the JSON payload of a backend event to a reusable function in App.js, parseBackendPayload, and used it in the new event handlers for rolling dice as well as updating loadSheetHandler.