Updating Tracker/Sheet components
Updated the Tracker component to take an optional function pointer as a property in TrackProps. This gives it similar generic versatility as was added to the Dots component.

Subsequently moved the updateHealth and updateWillpower state calls into the Sheet component, allowing the Tracker component to remain relatively generic.

Also updated both with a more measured use of references and smart pointers.