Made Size a Select element
Since Size needs to be variable, it needs to have a functional input, but to my knowledge, even though the game provides comparisons for sizes from 0 up to 100, the only sizes that characters can achieve permanently are 5 (default), 4 (Small-Framed merit), and 6 (Giant merit). So a Select element makes more sense than a text box that can contain any value.