Implemented removing Disciplines
Implemented the same method of removing Disciplines as is used for Merits.

Tweaked some wording in the Disciplines and Merits components.

Updated the styles of interactive elements to be a bit more dynamic. Things like the tracker box or dot "lighting up" when hovering the mouse over it and changing the cursor to a pointer.