Modified Attributes and Skills stateful data structures
Attributes and Skills are now stored, in the global state, as HashMaps keyed by BaseAttributeType and BaseSkillType respectively.

This allows the UI code to be much more concise, using iterators to generate most of the output.

It also allows for easier and more precise programmatic access to specific values in their global states.