Update disciplines.rs
Renamed DevotionField::Reference to Duration, as that is more immediately useful information to have on the sheet.

Also, added "& Powers" to the Devotions label since it fits well enough to store information about specific Discipline powers which have similar information associated with them as well, such as Protean's Unmarked Grave.