Abstracted Character state
Abstracted the common Chronicles of Darkness character state data into a BaseCharacter struct, which also implements the StatefulTemplate Trait. This will reduce the amount of boilerplate code needed when adding in more CoD templates.

Replaced the fields in vtr2e::templates::Kindred with one field of type BaseCharacter. Updated Kindred's push and pull methods to account for this change.

Also added the global state definitions for Aspirations, Beats, and Experience.