Fixed issue with File > New
After changing the Attributes and Skills into Vec's, the default BaseCharacter was being created with no attributes or skills, so the components were hitting a "no value at index" runtime error.

So, in addition to protecting against that runtime error in the future, I implemented the necessary custom default() functions to ensure the situation doesn't happen again.

The BaseCharacter struct was getting a little big so I moved it into its own module: cod::character

Also, added a validate() function to the StatefulTemplate trait to assist in preventing this problem when loading messy JSON data. The goal is to scan for, and repair, issues like this before trying to use the data.