Update app.rs
Using Dioxus' conditional "then" rendering pattern, I've landed on a way to both genericize the save/load methods and also actually use them properly.

It's not as clean a solution as I want but I need to do more research to determine if what I want to do is even possible.

Note: the file path that is loaded from or saved to is still hard coded. Need to build a file selector UI.