Documentation first pass
Added doc comments on most things, whether they really needed them or not.

Implemented Props structs for those components which were using inline-props, as the macro is not receptive to doc comments of any kind.

Renamed the cod::Base* structs/enums to cod::Core*, in preparation for adding in a Chronicles of Darkness Core sheet.

Unified the cod::components::Attributes and Skills components to use a single TraitProps struct.