Major Refactor
My original plan for organizing the code files was clearly starting to get a little out of hand. So I decided to move at least the non-UI code into a smaller number of files: enums/state/structs.

I also pulled more data out of the CoD game system states down into the CoD Core state, namely all the Advantages and all the Details. Characters all have mostly the same TYPES of advantages and details, even if the nomenclature is intentionally different, in every CoD game. This greatly reduces the amount of code required to get a new CoD sheet up and running as well as cleaning up the global state quite a bit.