Implemented core CoD Mortal
The CoreCharacter struct is essentially a Chronicles of Darkness Mortal, and there is a sheet for that in the core book.

Modified Details to account for the minor differences between Mortals and the rest of the CoD family.

Replaced Aspirations with SimpleEntryList. Added a Conditions list to CoreCharacter, also represented by a SimpleEntryList on the Mortal sheet.

Updated styles so everything renders correctly across all three sheets.