Improved DotEntryList
Improved DotEntryList to also be able to operate like the Disciplines component, provided a select box to add new entries from a static list and without the ability to change the name of each option. Expects a Vec<(String, usize)> for select options instead of a map.

Replaced the Disciplines component on vtr2e sheet.