Updated Serialization
Updated Load/Save functionality to store the game system along with the sheet itself. This means loading a sheet will automatically switch to the correct game system.

Created a new struct, SaveData, to achieve this. The simplest way to handle this is to first serialize the Sheet pulled down from the global state into a JSON String, put that into a SaveData alongside the current GameSystem, and then serialize the SaveData instance to file. Loading is the same process in reverse.

Removed the Game System menu, as it is now irrelevant. Replaced the File > New menu item with a submenu, so you select for which game system you want to create a new sheet.