Added Mage Spellcasting Component
Added a UI component for Mage: The Awakening, specifically designed to simply the process of determining the various factors that go into casting a spell.

The styling still needs some work, and some of the more intricate complexities of the system are still not accounted for, but it is at least mostly functional at this point.