Updated Spellcasting Calculator
The spellcasting calculator is mostly finished now. Finished up a couple bits of functionality, namely Rote casting benefiting from the sheet's current dots in the selected Skill when using the Mudra Yantra and selecting multiple Yantras easily.

I also added a couple buttons to make life easier. "Clear Yantras" resets the list of currently used Yantras. "Reset" clears the whole Spellcasting Calculator's local state, so you don't need to manually deselect anything when you want to cast a new spell.

Finally, refactored the structure and styles of the output so it looks... as good as it's going to for now.