Added Home Screen
Moved app.rs into components::core where it belongs.

Added home.rs and HomeScreen component. It is the default UI when no game system is selected and contains buttons to select a game system. This replaces the old File menu method of switching game systems: File > New now displays the HomeScreen component.

Added a None value to the GameSystem enumeration, to denote when to display the HomeScreen UI. Updated event handlers in App component to properly handle when GameSystem::None is the current game system.

Moved a couple styles out of Sheet.styl and into app.styl to accommodate an app level UI component that isn't marked with the .sheet CSS class.