Implemented Close Sheet Functionality
Implemented a closeActiveSheet() method in SheetManager to close the currently active tab in the SheetTabs container.

As I am unaware of a means of connecting to signals from the generated tab itself, for now I've added a File Menu item "Close Sheet" to handle this process. It seems that there is activity on this in Godot's github as I type though, so a better solution should be available in the near future.

Also, I decided to get rid of the FileMenu event handler methods for those events which only require a single line of code.