Added Skill Specialties
Created Skill.Specialty data structure.
Created SpecialtyList node, similar to ItemList but based on a Values property of type List<Skill.Specialty>. The generated inputs are an HBoxContainer, containing an OptionButton to select the associated Skill and a LineEdit to enter the specific specialty.
Added SpecialtyList to the Skill scene, below the skills.
Updated CoreSheetLogic.InitAndConnect() to handle SpecialtyList nodes.