Implementing Changeling: The Lost Sheet
First a big refactor of the *LogicSheet scripts, as well as Attributes and Skills collections, which I'm not going to take the time to move out into their own commit. But now CoreLogicSheet handles all of the common functions for those inputs that all Chronicles of Darkness sheets will possess. Also, greatly simplified the data structures where possible to avoid issues with Godot being able to marshall them in Signals.

Updated Merits tab to be Merits & Flaws. Added Flaws as a separate data structure within CodCore.

Added OCSM.TextEditUtilities class, largely just to implement an autosize method for TextEdits, rather than having to set a size and deal with scrolling within the node itself.

Started building the Contract scene which will be used in the ContractsList node. Incomplete, at the moment.

Removed several files that are no longer necessary.