Normalized Sheet and SheetData Types
Renamed the *SheetLogic.cs files to just *Sheet.cs.

Created CharacterSheet.cs as a base class for all *Sheet classes; Also contains the ICharacterSheet interface, used in SheetManager.cs to avoid needing to cast to a specific type to get the SheetData as a JSON string.

Sorted the methods in SheetManager.cs and added method getActiveSheetJsonData() to give a generic means to retrieve the current sheet's data as a json string, primarily for saving to file.