Implementing Metadata Infrastructure
Completed Add, Edit, and Delete functionality for Changeling: The Lost metadata. The sheet still needs to be wired up to let the user select Contracts, Merits, and Flaws rather than just typing them in, but apart from that it's fully functional for this sheet.

Removed the dynamically generated Add menu entry from MetadataMenu. Manage Metadata should display a popup that allows the user to add/edit/delete all the metadata for the current game system.