Implementing Metadata Infrastructure
Added MeritList as an extension of ItemDotsList for the CoD Merit type.

Replaced SignalPayload with the Transport class. Instead of trying to encapsulate entire lists, encapsulate individual objects and build a List<Transport<T>> to pass through signals.

Updated the Merits tab on ChangelingSheet to utilize MeritList as well as updated the addExistingMerit method to account for this change as well.