Namespaced Nodes
I decided to organize the nodes into namespaces just because it would inevitably become an issue as more and more game systems are added. The slight increase in inconvenience will be worth it in the end.

Renamed ItemList to EntryList to avoid naming collision with Godot.ItemList.

Also, started working on the D&D 5E sheet but not going to both fishing those changes out of this amalgamation.