Implementing D&D 5E Sheet
Set up the core infrastructure for adding and editing metadata for D&D 5E.

Updated MetadataMenu to open the D&D 5E AddEditMetadata scene.

Implemented the basic concrete form of a FeaturefulMetadataEntry node, BackgroundEntry. The UI for adding new Features isn't implemented yet but BackgroundEntry is ready to receive it.