Implementing D&D 5E Sheet
Renamed the Changeling and D&D 5E AddEditMetadata nodes to CodChangelingAddEditMetadata and DndFifthAddEditMetadata respectively. Need to remember to avoid naming collisions in the future just to keep things simple.

Added Features tab to DndFifthAddEditMetadata.

Added a FeatureOptionButton to the FeaturefulMetadataEntry base class. Allows adding and removing features to whatever featureful metadata is being manipulated. Functional on the BackgroundsEntry node.

Created displayDeleteConfirmation in NodeUtilities, since it was relatively easy to genericize and is definitely going to be reused in multiple places.

Renamed NewFeature to FeatureEntry, for clarity as to its purpose.

Updated the Default style with an entry for RichTextLabel now that it is being used. Further styling needs to occur for the display of Features and their FeatureSections, but it's functional for now.