Implementing D&D 5E Sheet
Added Class and Race to the metadata interface.

Updated Featureful to contain a list of FeatureSections, as Classes and Races generally have a considerable amount of flavor text associated with them. Backgrounds might as well, who knows? Updated associated data structures and nodes to account for this change.

Created ClassEntry, RaceEntry, ClassOptionsButton, and RaceOptionsButton.

Updated the main Fifth sheet to use RaceOptionsButton and BackgroundOptionsButton for the Race and Background, respectively.