Implementing D&D 5E Sheet
Updated metadata so that the various lists are sorted at appropriate times.

Added select(string text) method to ClassOptionsButton to facilitate setting the Selected index via matching the item's text.

Updated Feature node so that the Description is hidden when the details are toggled, as the Description is literal metadata meant to be a means of rapidly identifying what the feature is or does, while the Text of the feature is part of the detailed data to be displayed along with any FeatureSections that exist.