Updated FeaturefulMetadataEntry
Fixed FeaturefulMetadataEntry so, when displaying associated Features, the instances are updated AFTER being added to the Node tree, as the Feature.update() method depends upon Scene Unique Name searches.

Updated the Signal definition to state that a list of Feature data structures is being sent, not a list of Feature nodes.