Implementing D&D 5E Sheet
Started implementing the necessary data structures to handle a character's inventory. All items will either be InventoryItem, or extend it.

InventoryArmor, as may be expected, defines an item that can be equipped and provides an Armor Class value. Might end up changing the nomenclature at some point as "InventoryArmor" isn't sitting quite right with me, but it works for the moment.

Equipment is a struct to identify those items which are currently equipped on a character. Mostly useful for things like calculating Armor Class based on the currently equipped set of InventoryArmor.