Organizing DnD Nodes/Scripts
Moved all the *OptionsButton scripts into scripts/nodes/dnd/fifth/options as there are a lot already and likely to be even more in the future. Also, streamlined the loop syntax using the List.ForEach method instead of the full foreach() syntax.

Removed the Equipment scene and EquipmentNode script, as that whole interface is obsolete.