Expanding DnDFifthContainer
Renamed Armor property to Armors, for consistency.
Added Containers, Items, and Weapons for ItemContainer, Item, and ItemWeapon types respectively.

Added an AllItems getter property which returns a new list contains the contents of all the different item lists, for convenience. Emphasis on -new- list; changes to the list it returns will not be preserved in the metadata.

Updated references to DnDFifthContainer to account for the changes. Also further unification of foreach loops to the function form.