Updated ConfirmQuit Functionality
With the transition to popups extending Window, the functionality to limit clicking while the popup is displayed is built in now. Simplified ConfirmQuit to just be a ConfirmationDialog node that handles quitting or not quitting the application.

The popup automatically hides on cancel by default.
PopupCentered() ensures that it displays in the center of the window.