App and FileMenu Layout Issues
Resolved most, if not all, of the layout issues with the App scene.

Checked over the node properties for OpenSheet and SaveSheet scenes to ensure they're set up properly.

Fiddled with the layout of NewSheet scene to resolve the differently-sized-buttons issue and to ensure it's positioned correctly, not overlapping the menu.

Building the menu items in FileMenu in its _Ready() method since it provides a bit more nuanced control of each item.

Updated how the Quit menu item sends the close request notification. For some reason, even though it uses the exact same mechanism and I have confirmed the methods involved are being called, clicking the Quit menu item does not cause the ConfirmQuit dialog to pop up. However, using the OS close window button DOES cause the dialog to pop up. Using the EXACT SAME MECHANISM. Not sure what's going on there but it's small enough that I can move on for now.