Fixed ContractsList
The actual problem was the Attribute option button node wasn't set to display an empty option, so there were never any empty entries to be removed.

In addition, to facilitate this sort of operation in the future, created IEmptiable to normalize the ability to determine if an object is empty.
Implemented IEmptiable on OCSM.CoD.CtL.Contract.

Updated ContractsList to utilize Contract.Empty rather than using a gigantic adhoc logical statement.