Tweaked Node Spacing
Added MarginContainers to various things in the Changeling sheet to help with spacing things out so they feel good.
Modified the third column in ChangelingAdvantages to stretch out a bit more both to give more room for visible text in its LineEdits but also to compress the first and second columns as there was a little too much empty space around both of them.