OptionButton Refactor
Implemented some utility methods on OptionButton via C# Extension.
Updated all uses of OptionButton to utilize the new methods, eliminating a large amount of now unnecessary logic.

Refactored to utilize LINQ or even just .ForEach() for those loops which could be refactored to use it.

Updated the ToString() in OCSM.Die to use String.Format() with a constant format string.

Added a replaceItems() function to CustomOption which handles clearing the existing option items, adding an empty string option if necessary, adding the new list of option items, and preserving the previously selected index. This allows the refreshMetadata() overrides to have far less duplicate code as they are now solely focused on building the new list of option items.

Started refactoring the use of enums to rely more heavily on C#'s System.Enum static methods. The Label attribute is still useful for enum values whose human-readable name should differ from the value name but it is less necessary overall now. I'm sure there are a lot of enums in the project that still need to be reassessed but it's a start.