Redesign Metadata
Resolved error that was occurring when updating the metadata in any way; It had to do with how the metadata-driven options buttons refreshed their items and the need to manually disconnect custom signals.

Updated NodeUtilities.DisplayDeleteConfirmation to remove unused parameters.

Updated the use of NodeUtilities.DisplayDeleteConfirmation so the confirmation node is added as a child of the calling node; Resolves an issue with creating a new exclusive window.