Update FileSystemUtilities.cs
Using CommonApplicationData on non-Windows platforms requires elevated user rights. At the moment, it's not really worth asking the user to authenticate in order to share the metadata between multiple users. Just store it in the user's ApplicationData folder. Easy enough to change it later if this becomes a desirable feature.