Update Fifth.tscn, DndFifthAddEditMetadata.tscn, FeatureEntry.tscn, and 20 more files
Renamed Race to Species, since that is where 5E is heading with the new 2024 PHB.

Removed the Featureful type and UI elements; Design change such that Features associate themselves with various metadata rather than the metadata containing their own feature lists.

Updated metadata editing and use so that Backgrounds, Classes, and Species are all of the generic Metadata type rather than distinct types; Features now contain an instance of Tags and are now tagged to be associated with a background, a class, and/or a species; The UI will need to be updated to allow being tagged with more than one background/class/species at once.

Updated MetadataOption node with a SelectedMetadata property for easily getting and setting the node's current selected value as a Metadata instance, rather than needing to interact directly with the OptionButton API.