Update Fifth.tscn and DndFifthSheet.cs
Implemented dynamic maximum value for the MaxHP SpinBox based on the Constitution ability modifier and the hit dice and levels of all classes in the sheet.
When the MaxHP value is changed, the maximum value of the CurrentHP SpinBox is set to the new MaxHP value.
Removed the Hit Dice SpinBox from the main sheet interface, as the hit dice are now managed within the Classes tab.
Updated Inspiration label to be Heroic Inspiration, following the 2024 PHB change.