Update DynamicMetadataLabel.tscn, DynamicNumericLabel.tscn, DynamicTextLabel.tscn, and 4 more files
Added new "dynamic" label nodes. Essentially, the normal display is a Label node but when Actions.Click'ed the label is replaced with the appropriate input node. When the input node loses focus, the Label node is shown again with the updated value.

The Numeric variant has the potential for some oddities, just because of the nature of the SpinBox node and how it deals with Focus. But this is good enough for now.