Update implementation.rs, mod.rs, mode.rs, and 3 more files
Moved StatefulMode into its own file.
Updated StatefulTrack to use a composite template rather than "manually" creating the child widgets.
Updated StatefulTrack so that, in BoxOne mode, the fill image is set to SlashTwo instead of SlashOne.
Updated StatefulTrack so that, in CircleOne and BoxOne modes, clicking on one of its children sets the value to that button's index value rather than simply adding one to the current value.
Created a constant for the StateToggled signal name.