Update mod.rs, implementation.rs, implementation.rs, and 5 more files
Added constants to DotLabel for the default grid width and height.
The focus should end up on the last DotLabel in a DotLabelList after a refresh, though the user still needs to tab to get to the label entry.
Replaced static placeholders in the Changeling, Mage, and Vampire sheets with DotLabelList.